requirements.txt
# Core dependencies
asyncio>=3.4.3
aiohttp>=3.8.0
typing-extensions>=4.0.0

# Configuration and environment
pyyaml>=6.0
python-dotenv>=0.19.0

# Video processing
yt-dlp>=2023.3.4
pytube>=12.1.0
moviepy>=1.0.3

youtube-transcript-api>=0.6.1
isodate>=0.6.1

# YouTube API
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=0.4.0

# AI/NLP
openai>=1.0.0
numpy>=1.21.0
pandas>=1.4.0

# Web framework and UI
streamlit>=1.22.0

# Utilities
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0
pillow>=9.0.0

# Logging and monitoring
structlog>=22.1.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Development tools
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
